requirements.txt

openai
python-dotenv
slack-sdk